+++ src/argilla/client/feedback/dataset/local/mixins.py
@@ -12,13 +12,11 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from typing import TYPE_CHECKING, Dict, List, Literal, Optional, Type, Union
+from typing import TYPE_CHECKING, List, Literal, Optional, Type, Union
 from uuid import UUID
 
-from tqdm import trange
-
 from argilla.client.api import ArgillaSingleton
-from argilla.client.feedback.constants import PUSHING_BATCH_SIZE
+from argilla.client.feedback.dataset.helpers import get_dataset_by_name_and_workspace
 from argilla.client.feedback.dataset.remote.dataset import INCLUDE_ALL_VECTORS_PARAM, RemoteFeedbackDataset
 from argilla.client.feedback.schemas.enums import FieldTypes, MetadataPropertyTypes, QuestionTypes
 from argilla.client.feedback.schemas.fields import TextField
@@ -44,7 +42,6 @@ from argilla.client.feedback.schemas.remote.questions import (
 )
 from argilla.client.feedback.schemas.types import AllowedMetadataPropertyTypes
 from argilla.client.feedback.schemas.vector_settings import VectorSettings
-from argilla.client.feedback.utils import feedback_dataset_in_argilla
 from argilla.client.sdk.commons.errors import AlreadyExistsApiError
 from argilla.client.sdk.v1.datasets import api as datasets_api_v1
 from argilla.client.workspaces import Workspace
@@ -55,7 +52,6 @@ if TYPE_CHECKING:
     from argilla.client.client import Argilla as ArgillaClient
     from argilla.client.feedback.dataset.local import FeedbackDataset
     from argilla.client.feedback.dataset.local.dataset import FeedbackDataset
-    from argilla.client.feedback.schemas.records import FeedbackRecord
     from argilla.client.feedback.schemas.types import (
         AllowedFieldTypes,
         AllowedMetadataPropertyTypes,
@@ -70,6 +66,19 @@ if TYPE_CHECKING:
         FeedbackMetadataPropertyModel,
         FeedbackQuestionModel,
     )
+
+
+def _prepare_workspace(client, workspace):
+    if workspace is None:
+        workspace = client.get_workspace()
+
+    if workspace is None:
+        raise ValueError("No workspace provided and no active workspace found.")
+
+    if isinstance(workspace, str):
+        workspace = Workspace.from_name(workspace)
+
+    return workspace
 
 
 class ArgillaMixin:
@@ -203,35 +212,6 @@ class ArgillaMixin:
             ArgillaMixin.__delete_dataset(client=client, id=id)
             raise Exception(f"Failed while publishing the `FeedbackDataset` in Argilla with exception: {e}") from e
 
-    @staticmethod
-    def __push_records(
-        records: List["FeedbackRecord"],
-        client: "httpx.Client",
-        id: UUID,
-        question_name_to_id: Dict[str, UUID],
-        show_progress: bool = True,
-    ) -> None:
-        if len(records) == 0:
-            return
-
-        for i in trange(
-            0, len(records), PUSHING_BATCH_SIZE, desc="Pushing records to Argilla...", disable=show_progress
-        ):
-            try:
-                datasets_api_v1.add_records(
-                    client=client,
-                    id=id,
-                    records=[
-                        record.to_server_payload(question_name_to_id=question_name_to_id)
-                        for record in records[i : i + PUSHING_BATCH_SIZE]
-                    ],
-                )
-            except Exception as e:
-                ArgillaMixin.__delete_dataset(client=client, id=id)
-                raise Exception(
-                    f"Failed while adding the records to the `FeedbackDataset` in Argilla with exception: {e}"
-                ) from e
-
     def push_to_argilla(
         self: Union["FeedbackDataset", "ArgillaMixin"],
         name: str,
@@ -254,70 +234,50 @@ class ArgillaMixin:
         client: "ArgillaClient" = ArgillaSingleton.get()
         httpx_client: "httpx.Client" = client.http_client.httpx
 
-        if workspace is None:
-            workspace = Workspace.from_name(client.get_workspace())
-
-        if isinstance(workspace, str):
-            workspace = Workspace.from_name(workspace)
-
-        dataset = feedback_dataset_in_argilla(name=name, workspace=workspace)
-        if dataset is not None:
-            raise RuntimeError(
-                f"Dataset with name=`{name}` and workspace=`{workspace.name}` already exists in Argilla, please"
-                " choose another name and/or workspace."
-            )
+        workspace = _prepare_workspace(client, workspace)
+        created_dataset = _create_argilla_dataset_or_raise(httpx_client, name, workspace, dataset=self)
 
         try:
-            new_dataset: "FeedbackDatasetModel" = datasets_api_v1.create_dataset(
+            ArgillaMixin.__add_fields(fields=self.fields, client=httpx_client, id=created_dataset.id)
+            ArgillaMixin.__add_questions(questions=self.questions, client=httpx_client, id=created_dataset.id)
+
+            if self.metadata_properties:
+                ArgillaMixin.__add_metadata_properties(
+                    metadata_properties=self.metadata_properties, client=httpx_client, id=created_dataset.id
+                )
+
+            if self.vectors_settings:
+                ArgillaMixin.__add_vectors_settings(
+                    vectors_settings=self.vectors_settings, client=httpx_client, id=created_dataset.id
+                )
+
+            ArgillaMixin.__publish_dataset(client=httpx_client, id=created_dataset.id)
+
+            # TODO: Remote dataset should connect all settings by API calls requested on demand.
+            #  Once is done, this prefetch info should be removed.
+            fields = ArgillaMixin.__get_fields(client=httpx_client, id=created_dataset.id)
+            questions = ArgillaMixin.__get_questions(client=httpx_client, id=created_dataset.id)
+
+            remote_dataset = RemoteFeedbackDataset(
                 client=httpx_client,
+                id=created_dataset.id,
                 name=name,
-                workspace_id=workspace.id,
+                workspace=workspace,
+                created_at=created_dataset.inserted_at,
+                updated_at=created_dataset.updated_at,
+                fields=fields,
+                questions=questions,
                 guidelines=self.guidelines,
                 allow_extra_metadata=self.allow_extra_metadata,
-            ).parsed
-            argilla_id = new_dataset.id
-        except Exception as e:
-            raise Exception(f"Failed while creating the `FeedbackDataset` in Argilla with exception: {e}") from e
-
-        ArgillaMixin.__add_fields(fields=self.fields, client=httpx_client, id=argilla_id)
-        fields = ArgillaMixin.__get_fields(client=httpx_client, id=argilla_id)
-
-        ArgillaMixin.__add_questions(questions=self.questions, client=httpx_client, id=argilla_id)
-        questions = ArgillaMixin.__get_questions(client=httpx_client, id=argilla_id)
-        question_name_to_id = {question.name: question.id for question in questions}
-
-        if self.metadata_properties:
-            ArgillaMixin.__add_metadata_properties(
-                metadata_properties=self.metadata_properties, client=httpx_client, id=argilla_id
             )
 
-        if self.vectors_settings:
-            ArgillaMixin.__add_vectors_settings(
-                vectors_settings=self.vectors_settings, client=httpx_client, id=argilla_id
-            )
+            if len(self.records) > 0:
+                remote_dataset.add_records(self.records, show_progress)
 
-        ArgillaMixin.__publish_dataset(client=httpx_client, id=argilla_id)
-
-        ArgillaMixin.__push_records(
-            records=list(self.records),
-            client=httpx_client,
-            id=argilla_id,
-            show_progress=show_progress,
-            question_name_to_id=question_name_to_id,
-        )
-
-        return RemoteFeedbackDataset(
-            client=httpx_client,
-            id=argilla_id,
-            name=name,
-            workspace=workspace,
-            created_at=new_dataset.inserted_at,
-            updated_at=new_dataset.updated_at,
-            fields=fields,
-            questions=questions,
-            guidelines=self.guidelines,
-            allow_extra_metadata=self.allow_extra_metadata,
-        )
+            return remote_dataset
+        except Exception as ex:
+            ArgillaMixin.__delete_dataset(client=httpx_client, id=created_dataset.id)
+            raise ex
 
     @staticmethod
     def __get_fields(client: "httpx.Client", id: UUID) -> List["AllowedRemoteFieldTypes"]:
@@ -377,7 +337,7 @@ class ArgillaMixin:
         """
         httpx_client: "httpx.Client" = ArgillaSingleton.get().http_client.httpx
 
-        existing_dataset = feedback_dataset_in_argilla(name=name, workspace=workspace, id=id)
+        existing_dataset = get_dataset_by_name_and_workspace(name=name, workspace=workspace, id=id)
         if existing_dataset is None:
             raise ValueError(
                 f"Could not find a `FeedbackDataset` in Argilla with name='{name}'."
@@ -473,6 +433,29 @@ class ArgillaMixin:
         except Exception as e:
             ArgillaMixin.__delete_dataset(client=client, id=id)
             raise Exception(f"Failed adding vectors to the `FeedbackDataset` in Argilla with exception: {e}") from e
+
+
+def _create_argilla_dataset_or_raise(
+    httpx_client: "httpx.Client", name: str, workspace: "Workspace", dataset: "FeedbackDataset"
+) -> "FeedbackDatasetModel":
+    argilla_dataset = get_dataset_by_name_and_workspace(name=name, workspace=workspace)
+
+    if argilla_dataset is not None:
+        raise RuntimeError(
+            f"Dataset with name=`{name}` and workspace=`{workspace.name}` already exists in Argilla, please"
+            " choose another name and/or workspace."
+        )
+
+    try:
+        return datasets_api_v1.create_dataset(
+            client=httpx_client,
+            name=name,
+            workspace_id=workspace.id,
+            guidelines=dataset.guidelines,
+            allow_extra_metadata=dataset.allow_extra_metadata,
+        ).parsed
+    except Exception as e:
+        raise Exception(f"Failed while creating the `FeedbackDataset` in Argilla with exception: {e}") from e
 
 
 class TaskTemplateMixin: