+++ tests/unit/client/feedback/dataset/local/test_dataset.py
@@ -16,6 +16,7 @@ from typing import TYPE_CHECKING, List, Type
 
 import numpy.array_api
 import pytest
+from argilla import RatingQuestion
 from argilla.client.feedback.dataset.local.dataset import FeedbackDataset
 from argilla.client.feedback.schemas.fields import TextField
 from argilla.client.feedback.schemas.metadata import (
@@ -28,7 +29,11 @@ from argilla.client.feedback.schemas.records import FeedbackRecord
 from argilla.client.feedback.schemas.vector_settings import VectorSettings
 
 if TYPE_CHECKING:
-    from argilla.client.feedback.schemas.types import AllowedMetadataPropertyTypes
+    from argilla.client.feedback.schemas.types import (
+        AllowedFieldTypes,
+        AllowedMetadataPropertyTypes,
+        AllowedQuestionTypes,
+    )
 
 
 @pytest.mark.parametrize(
@@ -383,3 +388,153 @@ def test_not_implemented_methods():
         UserWarning, match="`filter_by` method is not supported for local datasets and won't take any effect. "
     ):
         assert dataset.filter_by() == dataset
+
+
+def test_init(
+    feedback_dataset_guidelines: str,
+    feedback_dataset_fields: List["AllowedFieldTypes"],
+    feedback_dataset_questions: List["AllowedQuestionTypes"],
+) -> None:
+    dataset = FeedbackDataset(
+        guidelines=feedback_dataset_guidelines,
+        fields=feedback_dataset_fields,
+        questions=feedback_dataset_questions,
+        allow_extra_metadata=False,
+    )
+
+    assert dataset.guidelines == feedback_dataset_guidelines
+    assert dataset.fields == feedback_dataset_fields
+    assert dataset.questions == feedback_dataset_questions
+    assert dataset.allow_extra_metadata == False
+
+
+def test_init_wrong_guidelines(
+    feedback_dataset_fields: List["AllowedFieldTypes"], feedback_dataset_questions: List["AllowedQuestionTypes"]
+) -> None:
+    with pytest.raises(TypeError, match="Expected `guidelines` to be"):
+        FeedbackDataset(
+            guidelines=[],
+            fields=feedback_dataset_fields,
+            questions=feedback_dataset_questions,
+        )
+    with pytest.raises(ValueError, match="Expected `guidelines` to be"):
+        FeedbackDataset(
+            guidelines="",
+            fields=feedback_dataset_fields,
+            questions=feedback_dataset_questions,
+        )
+
+
+def test_init_wrong_fields(
+    feedback_dataset_guidelines: str, feedback_dataset_questions: List["AllowedQuestionTypes"]
+) -> None:
+    with pytest.raises(TypeError, match="Expected `fields` to be a list"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=None,
+            questions=feedback_dataset_questions,
+        )
+    with pytest.raises(TypeError, match="Expected `fields` to be a list of `TextField`"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=[{"wrong": "field"}],
+            questions=feedback_dataset_questions,
+        )
+    with pytest.raises(ValueError, match="At least one field in `fields` must be required"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=[TextField(name="test", required=False)],
+            questions=feedback_dataset_questions,
+        )
+    with pytest.raises(ValueError, match="Expected `fields` to have unique names"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=[
+                TextField(name="test", required=True),
+                TextField(name="test", required=True),
+            ],
+            questions=feedback_dataset_questions,
+        )
+
+
+def test_init_wrong_questions(
+    feedback_dataset_guidelines: str, feedback_dataset_fields: List["AllowedFieldTypes"]
+) -> None:
+    with pytest.raises(TypeError, match="Expected `questions` to be a list, got"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=None,
+        )
+    with pytest.raises(
+        TypeError,
+        match="Expected `questions` to be a list of",
+    ):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=[{"wrong": "question"}],
+        )
+    with pytest.raises(ValueError, match="At least one question in `questions` must be required"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=[
+                TextQuestion(name="question-1", required=False),
+                RatingQuestion(name="question-2", values=[1, 2], required=False),
+            ],
+        )
+    with pytest.raises(ValueError, match="Expected `questions` to have unique names"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=[
+                TextQuestion(name="question-1", required=True),
+                TextQuestion(name="question-1", required=True),
+            ],
+        )
+
+
+def test_init_wrong_metadata_properties(
+    feedback_dataset_guidelines: str,
+    feedback_dataset_fields: List["AllowedFieldTypes"],
+    feedback_dataset_questions: List["AllowedQuestionTypes"],
+) -> None:
+    with pytest.raises(TypeError, match="Expected `metadata_properties` to be a list"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=feedback_dataset_questions,
+            metadata_properties=["wrong type"],
+        )
+    with pytest.raises(ValueError, match="Expected `metadata_properties` to have unique names"):
+        FeedbackDataset(
+            guidelines=feedback_dataset_guidelines,
+            fields=feedback_dataset_fields,
+            questions=feedback_dataset_questions,
+            metadata_properties=[
+                IntegerMetadataProperty(name="metadata-property-1", min=0, max=10),
+                IntegerMetadataProperty(name="metadata-property-1", min=0, max=10),
+            ],
+        )
+
+
+@pytest.mark.parametrize(
+    "metadata_property",
+    (
+        TermsMetadataProperty(name="terms-metadata-diff-name", values=["a", "b", "c"]),
+        IntegerMetadataProperty(name="int-metadata-diff-name", min=0, max=10),
+        FloatMetadataProperty(name="float-metadata-diff-name", min=0.0, max=10.0),
+    ),
+)
+def test__unique_metadata_property(metadata_property: "AllowedMetadataPropertyTypes") -> None:
+    dataset = FeedbackDataset(
+        fields=[TextField(name="required-field", required=True), TextField(name="optional-field", required=False)],
+        questions=[TextQuestion(name="question", required=True)],
+        metadata_properties=[
+            TermsMetadataProperty(name="terms-metadata", values=["a", "b", "c"]),
+            IntegerMetadataProperty(name="int-metadata", min=0, max=10),
+            FloatMetadataProperty(name="float-metadata", min=0.0, max=10.0),
+        ],
+    )
+    dataset._unique_metadata_property(metadata_property)